// src/services/central-api/generated/model/ctl-role-management.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PL Cental API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CtlRoleManagement
 */
export interface CtlRoleManagement {
    /**
     * 
     * @type {number}
     * @memberof CtlRoleManagement
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof CtlRoleManagement
     */
    'name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlRoleManagement
     */
    'createTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlRoleManagement
     */
    'createBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlRoleManagement
     */
    'updateTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlRoleManagement
     */
    'updateBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlRoleManagement
     */
    'deleteTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlRoleManagement
     */
    'deleteBy'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof CtlRoleManagement
     */
    'isDelete'?: boolean | null;
    /**
     * 
     * @type {boolean}
     * @memberof CtlRoleManagement
     */
    'isActive'?: boolean | null;
}